POSCP-51 - rename Pos to CardProcessing
